Style: Code CleanUp
