+++ app/src/main/java/com/example/geeksasaeng/Chatting/ChattingFragment.kt
@@ -1,6 +1,6 @@
 package com.example.geeksasaeng.Chatting
 
-import com.example.geeksasaeng.Base.BaseFragment
+import com.example.geeksasaeng.Utils.BaseFragment
 import com.example.geeksasaeng.databinding.FragmentChattingBinding
 
 class ChattingFragment: BaseFragment<FragmentChattingBinding>(FragmentChattingBinding::inflate) {
