+++ app/src/main/java/com/example/geeksasaeng/Community/CommunityFragment.kt
@@ -1,6 +1,6 @@
 package com.example.geeksasaeng.Community
 
-import com.example.geeksasaeng.Base.BaseFragment
+import com.example.geeksasaeng.Utils.BaseFragment
 import com.example.geeksasaeng.databinding.FragmentCommunityBinding
 
 class CommunityFragment: BaseFragment<FragmentCommunityBinding>(FragmentCommunityBinding::inflate) {
