+++ app/src/main/java/com/example/geeksasaeng/Home/CreateParty/CreatePartyFragment.kt
@@ -1,7 +1,7 @@
 package com.example.geeksasaeng.Home.CreateParty
 
 import android.util.Log
-import com.example.geeksasaeng.Base.BaseFragment
+import com.example.geeksasaeng.Utils.BaseFragment
 import com.example.geeksasaeng.databinding.FragmentCreatePartyBinding
