+++ app/src/main/java/com/example/geeksasaeng/Home/Delivery/BannerFragment.kt
@@ -1,6 +1,6 @@
 package com.example.geeksasaeng.Home.Delivery
 
-import com.example.geeksasaeng.Base.BaseFragment
+import com.example.geeksasaeng.Utils.BaseFragment
 import com.example.geeksasaeng.databinding.FragmentBannerBinding
 
 class BannerFragment(val imgRes: Int) : BaseFragment<FragmentBannerBinding>(FragmentBannerBinding::inflate) {
