+++ app/src/main/java/com/example/geeksasaeng/Home/Delivery/LookPartyFragment.kt
@@ -1,7 +1,7 @@
 package com.example.geeksasaeng.Home.Delivery
 
 import android.os.Parcelable
-import com.example.geeksasaeng.Base.BaseFragment
+import com.example.geeksasaeng.Utils.BaseFragment
 import com.example.geeksasaeng.R
 import com.example.geeksasaeng.databinding.FragmentLookPartyBinding
