+++ app/src/main/java/com/example/geeksasaeng/Home/Delivery/Retrofit/DeliveryPartyListResponse.kt
@@ -6,3 +6,18 @@ data class DeliveryPartyListResponse(
     val message: String,
     val result: List<DeliveryPartyResult>
 )
+
+data class DeliveryPartyResult(
+    val partyIdx: Int,
+    val hashTag: List<String>?,
+    val title: String,
+    val content: String,
+    val orderTime: String?,
+    val currentMatching: Int,
+    val maxMatching: Int,
+    val location: String?,
+    val matchingStatus: String?,
+    val createdAt: String,
+    val updatedAt: String,
+    val status: String
+)
