+++ app/src/main/java/com/example/geeksasaeng/Home/Helper/HelperFragment.kt
@@ -1,6 +1,6 @@
 package com.example.geeksasaeng.Home.Helper
 
-import com.example.geeksasaeng.Base.BaseFragment
+import com.example.geeksasaeng.Utils.BaseFragment
 import com.example.geeksasaeng.databinding.FragmentHelperBinding
 
 class HelperFragment: BaseFragment<FragmentHelperBinding>(FragmentHelperBinding::inflate) {
