+++ app/src/main/java/com/example/geeksasaeng/Home/Market/MarketFragment.kt
@@ -1,6 +1,6 @@
 package com.example.geeksasaeng.Home.Market
 
-import com.example.geeksasaeng.Base.BaseFragment
+import com.example.geeksasaeng.Utils.BaseFragment
 import com.example.geeksasaeng.databinding.FragmentMarketBinding
 
 class MarketFragment: BaseFragment<FragmentMarketBinding>(FragmentMarketBinding::inflate) {
