+++ app/src/main/java/com/example/geeksasaeng/MainActivity.kt
@@ -1,7 +1,6 @@
 package com.example.geeksasaeng
 
-import android.os.Bundle
-import com.example.geeksasaeng.Base.BaseActivity
+import com.example.geeksasaeng.Utils.BaseActivity
 import com.example.geeksasaeng.Chatting.ChattingFragment
 import com.example.geeksasaeng.Community.CommunityFragment
 import com.example.geeksasaeng.Home.HomeFragment
