+++ app/src/main/java/com/example/geeksasaeng/Profile/ProfileFragment.kt
@@ -1,6 +1,6 @@
 package com.example.geeksasaeng.Profile
 
-import com.example.geeksasaeng.Base.BaseFragment
+import com.example.geeksasaeng.Utils.BaseFragment
 import com.example.geeksasaeng.databinding.FragmentProfileBinding
 
 class ProfileFragment: BaseFragment<FragmentProfileBinding>(FragmentProfileBinding::inflate) {
