+++ app/src/main/java/com/example/geeksasaeng/Utils/BaseActivity.kt
@@ -1,4 +1,4 @@
-package com.example.geeksasaeng.Base
+package com.example.geeksasaeng.Utils
 
 import android.content.Intent
 import android.os.Bundle
