+++ app/src/main/java/com/example/geeksasaeng/Utils/ViewUtils.kt
@@ -1,4 +1,4 @@
-package com.example.geeksasaeng.Base
+package com.example.geeksasaeng.Utils
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
